Renames RemoveCartRuleFromOrder to DeleteCartRuleFromOrder
